% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeloops.R
\name{repeat_t}
\alias{repeat_t}
\title{Set up a \code{repeat} loop with progress reporting}
\usage{
repeat_t(each = 1L)
}
\arguments{
\item{each}{Print an update every \code{each} iterations. Defaults to \code{1}.}
}
\value{
A loop specification (class \code{tl_loop}) to be piped into \code{\link[=grapes-do]{\%do\%}}.
}
\description{
Describes a \code{repeat} loop to be run with \code{\link[=grapes-do]{\%do\%}}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{repeat_t(each = 10) \%do\% \{ ...; if (done) break \}
}\if{html}{\out{</div>}}

It behaves like a base \code{repeat} loop while printing the iteration count and
elapsed time. As with \code{repeat}, the body must contain a \code{break} to stop.
Parallel execution is not available for \code{repeat} loops.
}
\examples{
x <- 0
repeat_t(each = 10) \%do\% {
  x <- x + 1
  if (x >= 30) break
}
}
\seealso{
\code{\link[=grapes-do]{\%do\%}}, \code{\link[=for_t]{for_t()}}, \code{\link[=while_t]{while_t()}}
}
