% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeloops-package.R
\docType{package}
\name{timeloops-package}
\alias{timeloops}
\alias{timeloops-package}
\title{timeloops: Self-Reporting Loops for R}
\description{
Drop-in replacements for \code{for}, \code{while} and \code{repeat} that print how far a
loop has progressed, how long it has been running, and (for \code{for_t}) an
estimated time to completion. Useful for long-running loops where you want
to know whether to wait or go for a coffee.
}
\seealso{
\code{\link[=for_t]{for_t()}}, \code{\link[=while_t]{while_t()}}, \code{\link[=repeat_t]{repeat_t()}}
}
\author{
\strong{Maintainer}: Giovanni Tinervia \email{giovannitinervia9@gmail.com}

}
\keyword{internal}
