% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeloops.R
\name{while_t}
\alias{while_t}
\title{Set up a \code{while} loop with progress reporting}
\usage{
while_t(cond, each = 1L)
}
\arguments{
\item{cond}{The condition checked before each iteration.}

\item{each}{Print an update every \code{each} iterations. Defaults to \code{1}.}
}
\value{
A loop specification (class \code{tl_loop}) to be piped into \code{\link[=grapes-do]{\%do\%}}.
}
\description{
Describes a \code{while} loop to be run with \code{\link[=grapes-do]{\%do\%}}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{while_t(x < 100, each = 10) \%do\% \{ ... \}
}\if{html}{\out{</div>}}

It behaves like a base \code{while} loop while printing the iteration count and
elapsed time. No ETA is shown because the number of iterations is unknown in
advance. Parallel execution is not available for \code{while} loops.
}
\examples{
x <- 0
while_t(x < 30, each = 10) \%do\% {
  x <- x + 1
}
}
\seealso{
\code{\link[=grapes-do]{\%do\%}}, \code{\link[=for_t]{for_t()}}, \code{\link[=repeat_t]{repeat_t()}}
}
